revert reporter setting
