+++ src/reporters.ts
@@ -1,11 +1,11 @@
 import { DescribeNode, Hook, ItNode, RootNode } from "./nodes.ts";
+import { getAncestry } from "./nodes_util.ts";
 import { colors } from "./deps.ts";
 
 export type TestReporter = {
   reportStart: (node: RootNode) => void;
-  reportEnd: (node: RootNode) => void;
   reportHookError: (hook: Hook, error: unknown) => void;
-  reportCase: (node: ItNode) => void;
+  getTestCaseName: (node: ItNode) => string;
 };
 
 const log = console.log;
@@ -36,9 +36,13 @@ function formatNode(
   return str;
 }
 
-export class DenoReporter implements TestReporter {
-  reportEnd() {}
-  reportCase() {}
+export class Reporter implements TestReporter {
+  getTestCaseName(node: ItNode) {
+    const hierarchy = getAncestry(node)
+      .map((node: DescribeNode | ItNode) => colors.gray(node.headline));
+    hierarchy.push(colors.bold(node.headline));
+    return hierarchy.join(" > ");
+  }
 
   reportStart(node: RootNode) {
     log(formatNode(node));
