+++ tests/test_util.ts
@@ -2,10 +2,9 @@ import { TestReporter } from "../src/reporters.ts";
 import { ItNode, TestFunction } from "../src/nodes.ts";
 
 export class SilentReporter implements TestReporter {
+  getTestCaseName = () => "";
   reportStart() {}
-  reportEnd() {}
   reportHookError() {}
-  reportCase() {}
 }
 
 export async function silentTest(node: ItNode, wrappedFn: TestFunction) {
